Review, document, refactor.  Mostly on hostnames_test.go and GET checks
- started proper documentation of the test (`doc.go`)
- reviewed, documented and refactored `hostnames_test.go`
  - reorganized subtest names
  - added GET check before each createTest, to avoid flakiness
- GET checks review
  - Fixed bug  (string checks were not running at all)
  - Made string checks contextMap-aware (so the strings can be templated
    with context information)
  - Made String() method return fields always in the same order
- Other refactoring
  - Moved code around, to files where they made more sense
  - Moved tmp path creation to module's `init()`, as more than one test
    may depend on it
